EVAL/OAI/requirements.txt
accessible-pygments==0.0.4
aiohttp==3.8.4
aiolimiter==1.1.0
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attr==0.3.1
attrs==19.2.0
azure-core==1.27.1
azure-storage-blob==12.16.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==5.0.1
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.19
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
docutils==0.18.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
exceptiongroup==1.1.1
executing==1.2.0
expiringdict==1.2.2
fastjsonschema==2.17.1
filelock==3.12.2
fqdn==1.5.1
frozenlist==1.3.3
furo==2023.5.20
google-api-core==2.11.0
google-auth==2.14.1
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==2.0.2
grpc-google-iam-v1==0.12.4
grpcio==1.56.0
grpcio-status==1.48.2
htmlmin==0.1.12
idna==3.4
ijson==3.2.2
imagesize==1.4.1
importlib-metadata==6.7.0
inflection==0.5.1
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-book==0.15.1
jupyter-cache==0.6.1
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_myst==2.0.0
jupyterlab_server==2.23.0
jupytext==1.14.7
label-studio==1.8.0
label-studio-converter==0.0.53
label-studio-tools==0.0.2
latexcodec==2.0.1
launchdarkly-server-sdk==7.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==3.0.1
multidict==6.0.4
myst-nb==0.17.2
myst-parser==0.18.1
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.6.7
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
openapi-codec==1.3.2
ordered-set==4.0.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdflatex==0.1.3
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2023.7.1
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybtex==0.24.0
pybtex-docutils==1.0.2
pycparser==2.21
pydantic==1.10.10
pydata-sphinx-theme==0.13.3
pyee==8.2.2
Pygments==2.15.1
pyoai==2.5.0
pyppeteer==1.0.2
pyRFC3339==1.1
pyrsistent==0.19.3
python-dateutil==2.8.1
python-json-logger==2.0.4
pytz==2022.7.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==3.5.3
regex==2023.6.3
requests==2.28.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
semver==2.13.0
Send2Trash==1.8.2
sentry-sdk==1.26.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==5.0.2
sphinx-basic-ng==1.0.0b1
sphinx-book-theme==1.0.1
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx-jupyterbook-latex==0.5.2
sphinx-mermaid==0.0.8
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.2.1
sphinx-togglebutton==0.3.2
sphinx_design==0.3.0
sphinx_external_toc==0.3.1
sphinx_inline_tabs==2023.4.21
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-imagehelper==1.1.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-svg2pdfconverter==1.2.2
SQLAlchemy==2.0.17
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.0
ua-parser==0.16.1
uc-micro-py==1.0.2
ujson==5.8.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
widgetsnbextension==4.0.7
xmljson==0.2.0
yarl==1.9.2
zipp==3.15.0